- refactor action and move different functionalities in better classes - merge releaseNotes and releaseNotesBuilder - update testcases to cover new structure
